requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
scikit-image>=0.19.0
scipy>=1.7.0
transformers>=4.20.0
tqdm>=4.64.0
openai>=1.40.0
PyMuPDF
# potrace
blend-modes
requests
google-genai